refactor: split host_boundary.go into 4 focused files and remove thin wrappers
